+++ app/src/main/java/org/simple/clinic/main/TheActivityComponent.kt
@@ -46,9 +46,9 @@ import org.simple.clinic.instantsearch.InstantSearchScreen
 import org.simple.clinic.login.applock.AppLockScreen
 import org.simple.clinic.login.applock.ConfirmResetPinDialog
 import org.simple.clinic.medicalhistory.newentry.NewMedicalHistoryScreen
-import org.simple.clinic.monthlyscreeningreports.complete.MonthlyScreeningReportCompleteScreen
-import org.simple.clinic.monthlyscreeningreports.form.QuestionnaireEntryScreen
-import org.simple.clinic.monthlyscreeningreports.list.MonthlyScreeningReportListScreen
+import org.simple.clinic.monthlyreports.complete.MonthlyReportCompleteScreen
+import org.simple.clinic.monthlyreports.form.QuestionnaireEntryScreen
+import org.simple.clinic.monthlyreports.list.MonthlyReportsScreen
 import org.simple.clinic.navigation.di.FragmentScreenKeyModule
 import org.simple.clinic.navigation.v2.Router
 import org.simple.clinic.navigation.v2.ScreenResultBus
@@ -164,8 +164,8 @@ interface TheActivityComponent :
     SelectLineListFormatDialog.Injector,
     QuestionnaireEntryScreen.Injector,
     PatientsTabLinkView.Injector,
-    MonthlyScreeningReportListScreen.Injector,
-    MonthlyScreeningReportCompleteScreen.Injector {
+    MonthlyReportsScreen.Injector,
+    MonthlyReportCompleteScreen.Injector {
   fun inject(target: TheActivity)
 
   @Subcomponent.Factory